+++ src/leiningen/new/db.clj
@@ -11,13 +11,13 @@
     (some #{"+sqlite"} features) :sqlite))
 
 (defn db-dependencies [options]
-  (into [['luminus-migrations "0.6.5"]
-         ['conman "0.8.3"]]
-        ({:postgres [['org.postgresql/postgresql "42.2.6"]]
-          :mysql    [['mysql/mysql-connector-java "8.0.16"]
+  (into [['luminus-migrations "0.6.6"]
+         ['conman "0.8.4"]]
+        ({:postgres [['org.postgresql/postgresql "42.2.8"]]
+          :mysql    [['mysql/mysql-connector-java "8.0.18"]
                      ['com.google.protobuf/protobuf-java "3.8.0"]]
-          :h2       [['com.h2database/h2 "1.4.199"]]
-          :sqlite   [['org.xerial/sqlite-jdbc "3.25.2"]]}
+          :h2       [['com.h2database/h2 "1.4.200"]]
+          :sqlite   [['org.xerial/sqlite-jdbc "3.28.0"]]}
          (select-db options))))
 
 (defn db-url [{:keys [sanitized] :as options} suffix]